refactor: EventInputForm 분리 (wip)
