+++ src/components/EventInputForm.tsx
@@ -1,0 +1,244 @@
+import {
+  Button,
+  Checkbox,
+  FormControl,
+  FormLabel,
+  Heading,
+  HStack,
+  Input,
+  Select,
+  Tooltip,
+  useToast,
+  VStack,
+} from '@chakra-ui/react';
+import React from 'react';
+
+import { useEventFormStore } from '../store/useEventFormStore';
+import { Event, EventForm, RepeatType } from '../types';
+import { findOverlappingEvents } from '../utils/eventOverlap';
+import { getTimeErrorMessage } from '../utils/timeValidation';
+
+const categories = ['업무', '개인', '가족', '기타'];
+
+// TODO: 나중에 정리
+type Props = {
+  events: Event[];
+  setOverlappingEvents: React.Dispatch<React.SetStateAction<Event[]>>;
+  setIsOverlapDialogOpen: React.Dispatch<React.SetStateAction<boolean>>;
+  saveEvent: (eventData: Event | EventForm) => Promise<void>;
+  notificationOptions: {
+    value: number;
+    label: string;
+  }[];
+};
+
+export const EventInputForm = ({
+  events,
+  saveEvent,
+  setIsOverlapDialogOpen,
+  setOverlappingEvents,
+  notificationOptions,
+}: Props) => {
+  const {
+    eventForm: {
+      title,
+      date,
+      startTime,
+      endTime,
+      description,
+      location,
+      category,
+      isRepeating,
+      repeat: { type: repeatType, interval: repeatInterval, endDate: repeatEndDate },
+      notificationTime,
+    },
+    setEventForm,
+    timeErrorRecord: { startTimeError, endTimeError },
+    editingEvent,
+    changeStartTime,
+    changeEndTime,
+    resetForm,
+  } = useEventFormStore();
+
+  const toast = useToast();
+
+  const addOrUpdateEvent = async () => {
+    if (!title || !date || !startTime || !endTime) {
+      toast({
+        title: '필수 정보를 모두 입력해주세요.',
+        status: 'error',
+        duration: 3000,
+        isClosable: true,
+      });
+      return;
+    }
+
+    if (startTimeError || endTimeError) {
+      toast({
+        title: '시간 설정을 확인해주세요.',
+        status: 'error',
+        duration: 3000,
+        isClosable: true,
+      });
+      return;
+    }
+
+    const eventData: Event | EventForm = {
+      id: editingEvent ? editingEvent.id : undefined,
+      title,
+      date,
+      startTime,
+      endTime,
+      description,
+      location,
+      category,
+      repeat: {
+        type: isRepeating ? repeatType : 'none',
+        interval: repeatInterval,
+        endDate: repeatEndDate || undefined,
+      },
+      notificationTime,
+    };
+
+    const overlapping = findOverlappingEvents(eventData, events);
+    if (overlapping.length > 0) {
+      setOverlappingEvents(overlapping);
+      setIsOverlapDialogOpen(true);
+    } else {
+      await saveEvent(eventData);
+      resetForm();
+    }
+  };
+
+  return (
+    <VStack w="400px" spacing={5} align="stretch">
+      <Heading>{editingEvent ? '일정 수정' : '일정 추가'}</Heading>
+
+      <FormControl>
+        <FormLabel>제목</FormLabel>
+        <Input value={title} onChange={(e) => setEventForm({ title: e.target.value })} />
+      </FormControl>
+
+      <FormControl>
+        <FormLabel>날짜</FormLabel>
+        <Input type="date" value={date} onChange={(e) => setEventForm({ date: e.target.value })} />
+      </FormControl>
+
+      <HStack width="100%">
+        <FormControl>
+          <FormLabel>시작 시간</FormLabel>
+          <Tooltip label={startTimeError} isOpen={!!startTimeError} placement="top">
+            <Input
+              type="time"
+              value={startTime}
+              onChange={(e) => changeStartTime(e.target.value)}
+              onBlur={() => getTimeErrorMessage(startTime, endTime)}
+              isInvalid={!!startTimeError}
+            />
+          </Tooltip>
+        </FormControl>
+        <FormControl>
+          <FormLabel>종료 시간</FormLabel>
+          <Tooltip label={endTimeError} isOpen={!!endTimeError} placement="top">
+            <Input
+              type="time"
+              value={endTime}
+              onChange={(e) => changeEndTime(e.target.value)}
+              onBlur={() => getTimeErrorMessage(startTime, endTime)}
+              isInvalid={!!endTimeError}
+            />
+          </Tooltip>
+        </FormControl>
+      </HStack>
+
+      <FormControl>
+        <FormLabel>설명</FormLabel>
+        <Input
+          value={description}
+          onChange={(e) => setEventForm({ description: e.target.value })}
+        />
+      </FormControl>
+
+      <FormControl>
+        <FormLabel>위치</FormLabel>
+        <Input value={location} onChange={(e) => setEventForm({ location: e.target.value })} />
+      </FormControl>
+
+      <FormControl>
+        <FormLabel>카테고리</FormLabel>
+        <Select value={category} onChange={(e) => setEventForm({ category: e.target.value })}>
+          <option value="">카테고리 선택</option>
+          {categories.map((cat) => (
+            <option key={cat} value={cat}>
+              {cat}
+            </option>
+          ))}
+        </Select>
+      </FormControl>
+
+      <FormControl>
+        <FormLabel>반복 설정</FormLabel>
+        <Checkbox
+          isChecked={isRepeating}
+          onChange={(e) => setEventForm({ isRepeating: e.target.checked })}
+        >
+          반복 일정
+        </Checkbox>
+      </FormControl>
+
+      <FormControl>
+        <FormLabel>알림 설정</FormLabel>
+        <Select
+          value={notificationTime}
+          onChange={(e) => setEventForm({ notificationTime: Number(e.target.value) })}
+        >
+          {notificationOptions.map((option) => (
+            <option key={option.value} value={option.value}>
+              {option.label}
+            </option>
+          ))}
+        </Select>
+      </FormControl>
+
+      {isRepeating && (
+        <VStack width="100%">
+          <FormControl>
+            <FormLabel>반복 유형</FormLabel>
+            <Select
+              value={repeatType}
+              onChange={(e) => setEventForm({ repeat: { type: e.target.value as RepeatType } })}
+            >
+              <option value="daily">매일</option>
+              <option value="weekly">매주</option>
+              <option value="monthly">매월</option>
+              <option value="yearly">매년</option>
+            </Select>
+          </FormControl>
+          <HStack width="100%">
+            <FormControl>
+              <FormLabel>반복 간격</FormLabel>
+              <Input
+                type="number"
+                value={repeatInterval}
+                onChange={(e) => setEventForm({ repeat: { interval: Number(e.target.value) } })}
+                min={1}
+              />
+            </FormControl>
+            <FormControl>
+              <FormLabel>반복 종료일</FormLabel>
+              <Input
+                type="date"
+                value={repeatEndDate}
+                onChange={(e) => setEventForm({ repeat: { endDate: e.target.value } })}
+              />
+            </FormControl>
+          </HStack>
+        </VStack>
+      )}
+
+      <Button data-testid="event-submit-button" onClick={addOrUpdateEvent} colorScheme="blue">
+        {editingEvent ? '일정 수정' : '일정 추가'}
+      </Button>
+    </VStack>
+  );
+};
